requirements/base.txt
Django==3.2.12
django-environ==0.9.0
psycopg2==2.9.3
djangorestframework==3.13.1

celery==5.2.3
flower==1.2.0
django-celery-results==2.4.0
django-celery-beat==2.2.1

sendbox-sdk==0.1.1

django-safedelete==1.3.0

whitenoise==6.2.0

django-filter==21.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-leaflet==0.28.3
django-storages==1.13.1
boto3==1.24.71
botocore==1.27.71

django-phonenumber-field[phonenumbers]==6.1.0
drf-jwt==1.19.2
rules==3.3.0
geopy==2.2.0
geos==0.2.3
djangorestframework-gis==1.0
numpy==1.23.2
firebase_admin==4.3.0
django_simple_history==3.2.0
django-import-export==3.0.1